Remove the widget code
This code hasn't seen any action since it was created in early 2010. It
has suffered major bitrot. The state of the code makes it impossible to
actually integrate into any modern website. The fact that the widget
requires running your own instance of OpenFlights is also a major
dealbreaker.

The demo as seen on the actual website is broken in a number of ways:
https://openflights.org/widget/demo

Closes https://github.com/jpatokal/openflights/issues/1224.